Show a 'waiting' icon until a 'watch' is activated (#245).
The Watch procedure may take some seconds and sometimes it is
confusing and user may click again and again. We're now showing
a 'loading' icon as an intermediate icon until the final stage
in order to prevent the user from multiple clicks.

The new 'spinner' animated GIF can be used in the future for
operations which take some time to finish (CSS class: ".waiting").