Added support for svn https repositories.
So far Tx included commit support only for repositories over SSH.
This patch adds support for https repos, by allowing the admin to
define the username and password for each 2nd-level domain he'd like
to serve. For example, we group together authentication for
'*.example.com'.

Authentication happens only for write operations with a decorator.

If the domain grouping proves inadequate in the future (eg. because
of '.ac.uk' domains, we'll re-think the credential grouping then.