Overhaul of the ``guess_language`` method using regexxes.
Should also address tickets #541, #542.

Moved the regexxes outside of the methods and compiled them once.
Thanks to Nikos Korkakakis for tips.

Notice the list comprehension in lib/types/pot.py: We used to have a
generator there '()'. But because this 'list' needs to be itereatable
for every call of it, having it as a generator will fail, because after
the first successful run, the generator won't regenerate/rewind.
Obviously a list is much more efficient since it will only be
calculated once.

The new method takes 0.1 millisecond more when hitting 'Refresh Stats'
and for the 'txstatsfresh' on Txn, when run against 30K files, it takes
0.30s, compared to 0.14s it took before. This is a penalty of appx.
0.01%.