Show messages under the correct translation box.
When the translator hit tab while translating, the focus would change to the
next translation box. So, any error messages were shown under that. This patch
changes that, so that lotte uses now the id of the original translation box the
request was originated from to display any error messages under the correct
translation bix.
This fixes bug 754. The patch was based on rtnpro's one at trac.