Improve lotte performance when filtering strings.
We now use specialized queries for each possible combination of filters. This
allows us eg to return an empty set immediately, if that would be the result.
Additionally, the perfomance of 'translated'/'untranslated' filters has been
improved by working around the former complex queries that postgres neeeded a
lot of time to process.