Make the source language used in a project a property of it.
Every resource in a project must use the same source language. So we define a
property for a project object to return the source language used.

There are two versions defined: one that returns a Language object and another
one that returns just the id of the source language (for efficiency reasons).