Fix: Handle case of empty last_commiter for statistics with zero translated strings.
In the case that a translation has never been modified, the variable
RLStats.last_committer will be None. In this case, access to the username of
the last_committer would result in an exception being raised. This was done in
the new API for getting the statistics of a resource.

This fix returns an empty string in the above described case.