Make Translation model have a foreign key to resources.
Each translation entry belongs to one resource in the course of its life, ie the data is read-only. Thus, it is safe to denormalize the resource.  So, instead of JOINing all the time with the relevant source entity, we just save the resource id along side the translation objects.

The patch includes migrations to add the necessary, column, fill in the
resource_id values and make the column NOT NULL from now on. The three
migrations *must* be run together.