Fix: Allow Project.objects.delete() overcome limitations in the handlers of releases.
The order that projects, resources, releases are deleted in such a case varies.
Semantically, this is correct, since django creates the foreign key constraints
with the specification "INITIALLY DEFERRED". SO, if we see that a resource has
no project, we can be sure that the project just got deleted and the releases
will be deleted, too.