Don't update PO-Revision-Date for every fetch
So far, whenever an empty translation (or rather, one that hadn't
ever been translated before) was fetched, the PO-Revision-Date
would be updated to the current time. Instead, we're now using the
time of creation for that translation. This means that when new
translations are fetched, the PO-Revision-Date won't change without
any actual new translations anymore.

This fixes bug 786 ( http://trac.transifex.org/ticket/786 ).