Optimize queries for translated, untranslated and user-translated strings.
We use the resource_id column as well for filtering. This allows the database
to reduce the search space for the strings resulting in much faster response
times.

This helps to avoid issues that show up, when there is a large list of source
strings in the IN clause of the query.