Performance fix: Fetch only needed columns in the access control form for projects.
Previous code resulted in using external memory to make the ORDER BY,
which meant a lot of seconds were spent on sorting. The change instructs
django's ORM to only fetch the id and name columns (those used in teh select
tag in the form and as a result the sorting is much faster. The new code is a
more than 20 times faster.