Added lotte_delete_translation signal and moved lotte signal under projects app
There seems to be a race condition while loading the addons that in some cases
prevents some signals to be properly registered. Moving the Lotte signals to
the project app namespace fixed the problem. Whoever this is not a definitive
solution. A FIXME was added to remember it do be done in the future.